# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Runs Service",
    "description": "REST API for tracking service execution runs and their associated costs. Supports hierarchical runs (parent-child), cost aggregation, and multi-tenant isolation via organizations.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authenticating requests"
      }
    },
    "schemas": {
...
      "Run": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organizationId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "brandId": {
            "type": "string",
            "nullable": true
          },
          "campaignId": {
            "type": "string",
            "nullable": true
          },
          "workflowName": {
            "type": "string",
            "nullable": true
          },
          "serviceName": {
            "type": "string"
          },
          "taskName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "parentRunId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "organizationId",
          "userId",
          "brandId",
          "campaignId",
          "workflowName",
          "serviceName",
          "taskName",
          "status",
          "parentRunId",
          "startedAt",
          "completedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "nullable": true
          }
        },
        "required": [
          "error"
        ]
      },
      "CreateRunRequest": {
        "type": "object",
        "properties": {
          "brandId": {
            "type": "string",
            "minLength": 1,
            "deprecated": true,
            "description": "Deprecated: use x-brand-id header instead. Kept for backwards compatibility; header takes precedence."
          },
          "campaignId": {
            "type": "string",
            "minLength": 1,
            "deprecated": true,
            "description": "Deprecated: use x-campaign-id header instead. Kept for backwards compatibility; header takes precedence."
          },
          "workflowName": {
            "type": "string",
            "minLength": 1,
            "deprecated": true,
            "description": "Deprecated: use x-workflow-name header instead. Kept for backwards compatibility; header takes precedence."
          },
          "serviceName": {
            "type": "string",
            "minLength": 1
          },
          "taskName": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "serviceName",
          "taskName"
        ]
      },
...
      "Cost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "runId": {
            "type": "string",
            "format": "uuid"
          },
          "costName": {
            "type": "string"
          },
          "costSource": {
            "$ref": "#/components/schemas/CostSource"
          },
          "quantity": {
            "type": "string"
          },
          "unitCostInUsdCents": {
            "type": "string"
          },
          "totalCostInUsdCents": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CostStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "runId",
          "costName",
          "costSource",
          "quantity",
          "unitCostInUsdCents",
          "totalCostInUsdCents",
          "status",
          "createdAt"
        ]
      },
      "CostSource": {
        "type": "string",
        "enum": [
          "platform",
          "org"
        ]
      },
      "CostStatus": {
        "type": "string",
        "enum": [
          "actual",
          "provisioned",
          "cancelled"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "AddCostsResponse": {
        "type": "object",
        "properties": {
          "costs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cost"
            }
          }
        },
        "required": [
          "costs"
        ]
      },
      "AddCostsRequest": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CostItem"
            },
            "minItems": 1
          }
        },
        "required": [
          "items"
        ]
      },
      "CostItem": {
        "type": "object",
        "properties": {
          "costName": {
            "type": "string",
            "minLength": 1
          },
          "costSource": {
            "$ref": "#/components/schemas/CostSource"
          },
          "quantity": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "status": {
            "$ref": "#/components/schemas/CostStatus"
          }
        },
        "required": [
          "costName",
          "costSource",
          "quantity"
        ]
      },
...
      "UpdateRunRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "failed"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/platform-runs": {
      "post": {
        "summary": "Create a platform-level run",
        "description": "Creates a run for a platform-level system operation (no org or user). Requires x-service-name header to identify the calling service. Field resolution priority: header > body (deprecated).",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Brand identifier, injected by workflow-service"
            },
            "required": false,
            "description": "Brand identifier, injected by workflow-service",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign identifier, injected by workflow-service"
            },
            "required": false,
            "description": "Campaign identifier, injected by workflow-service",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow name, injected by workflow-service"
            },
            "required": false,
            "description": "Workflow name, injected by workflow-service",
            "name": "x-workflow-name",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Platform run created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or missing x-service-name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/platform-runs/{id}/costs": {
      "post": {
        "summary": "Add costs to a platform run",
        "description": "Adds cost line items to a platform-level run. Unit costs are resolved automatically from the costs-service.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCostsRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Costs added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddCostsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unknown cost name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/platform-runs/{id}": {
      "patch": {
        "summary": "Update platform run status",
        "description": "Updates a platform-level run status to completed or failed. Sets completedAt automatically.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Run updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}